+++ compile_shader.sh
@@ -1,2 +1,2 @@
-glslc VulkanTutorial/shaders/shader.vert -o bin/vert.spv
-glslc VulkanTutorial/shaders/shader.frag -o bin/frag.spv
+glslc shaders/shader.vert -o bin/vert.spv
+glslc shaders/shader.frag -o bin/frag.spv
